+++ rmp/Cargo.toml
@@ -11,3 +11,4 @@ keywords = ["msgpack", "MessagePack"]
 
 [dependencies]
 byteorder = "^0.5"
+# num-traits = "^0.1"
