+++ rmp/tests/func/decode/mod.rs
@@ -1,14 +1,16 @@
 // Low-level interface.
 mod null;
 mod bool;
-mod int;
+mod sint;
+mod uint;
+// mod int;
 // mod float;
 // mod string;
 // mod bin;
 // mod array;
 // mod map;
 // mod ext;
-//
+
 // // High-level deserialize interface.
 // mod value;
 // mod value_ref;
